controlmapper/requirements.txt
amqp==5.0.6
asgiref==3.3.4
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.2.3
django-cors-headers==3.7.0
django-filter==2.4.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
gunicorn==20.1.0
importlib-metadata==4.0.1
kombu==5.1.0
lxml==4.6.3
Markdown==3.3.4
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
PyJWT==1.7.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
uWSGI==2.0.19.1
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.1
